Add date and time functionality
The user would benefits from being able to
enter dates in a computer-parsable format, as
it would allow for increased tracking capabilities.

Let's add date and time functionality to the bot,
as well as a command that lists out tasks with
dates matching with the date provided by the user.